% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-attributes.R
\name{as_attributes}
\alias{as_attributes}
\title{R objects as OpenTelemetry attributes}
\usage{
as_attributes(x)
}
\arguments{
\item{x}{A list of R objects, to be used as OpenTelemetry attributes.}
}
\value{
A named list that can be used as the \code{attributes} argument to
\code{\link[=start_span]{start_span()}}, \code{\link[=log]{log()}}, \code{\link[=counter_add]{counter_add()}}, etc.

If \code{x} is not named, or some names are the empty string or \code{NA}, then
integer numbers as used for the missing or invalid names.

If some elements in \code{x} are not of the natively supported R types in
OpenTelemetry (character, logical, double, integer), then
their printed form is captured using \code{\link[utils:capture.output]{utils::capture.output()}}.
\subsection{Limits}{

The number of attributes can be limited with the
OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT environment variable. The default is
100.

The length of the each attribute (vector) can be limited with the
OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT environment variable. The default is
100. Note that this is applied to the
length of each attribute as an R vector. E.g. it does \emph{not} currently
limit the number of characters in individual strings.
}
}
\description{
Convert a list of R objects to a form that is suitable as OpenTelemetry
attributes.
}
