% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_current_baggage_value}
\alias{set_current_baggage_value}
\alias{get_current_baggage_value}
\alias{remove_current_baggage_value}
\alias{get_current_baggage}
\alias{clear_current_baggage}
\title{Baggage API}
\usage{
set_current_baggage_value(key, value)

get_current_baggage_value(key)

remove_current_baggage_value(key)

get_current_baggage()

clear_current_baggage()
}
\arguments{
\item{key}{String (character scalar), the baggage key to set or query.}

\item{value}{String (character scalar), the baggage value to set. It
cannot be \code{NA}.}
}
\value{
\code{set_current_baggage_value()} returns \code{NULL}, invisibly.

\code{get_current_baggage_value()} returns a string scalar, the
value that belongs to \code{key}, or \code{NA_character_} if \code{key} is not set.

\code{remove_current_baggage_value()} returns \code{NULL}, invisibly.

\code{get_current_baggage()} returns a named character vector.

\code{clear_current_baggage()} returns \code{NULL}, invisibly.
}
\description{
Custom contextual information.
}
\details{
In OpenTelemetry, Baggage is custom contextual information that resides
next to context and is passed to other in-process contexts, and can be
passed to other processes as well.

Baggage is a key-value store, where both the keys and the values are
UTF-8 strings.

\code{set_current_baggage_value()} sets \code{value} for \code{key} in the
\emph{current context}. If \code{key} was previously set, it is overwritten.

\code{get_current_baggage_value()} queries the baggage value of \code{key} in the
\emph{current context}. You can use it to query values passed on from other
contexts, e.g. from parent spans, or from other processes. It returns
\code{NA_character_} if \code{key} is not set in the current context.

\code{remove_current_baggage_value()} removed \code{key} from the baggage of the
\emph{current context}, if it was set. If it was not set then it does not do
anything.

\code{get_current_baggage()} returns all keys and thei corresponding values
of the baggage of the \emph{current context}.

\code{clear_current_baggage()} removed all keys from the baggage of the
\emph{current context}.
}
