% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-name.R
\name{default_tracer_name}
\alias{default_tracer_name}
\alias{otel_tracer_name}
\title{Default tracer (and meter and logger) name for an R package}
\usage{
default_tracer_name(name = NULL)
}
\arguments{
\item{name}{Custom tracer name. If \code{NULL} then otel will construct a
a tracer (meter, logger) name according to the algorithm detailed
below.

TODO}
}
\description{
Default tracer (and meter and logger) name for an R package
}
