% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_active_span}
\alias{get_active_span}
\title{Returns the active span, if any}
\usage{
get_active_span(tracer = NULL)
}
\arguments{
\item{tracer}{Tracer object (\link{otel_tracer}) or tracer name to use.
If \code{NULL}, then otel uses an internal tracer. You usually do not
need to set this, the active span does not depend on the tracer.

Passing a tracer might give the wrong result. If it is a no-op
tracer, then the result is an invalid span (context), or no HTTP
headers, even if there is an active span. This happens if the
tracer's scope is turned off, e.g. via the \code{OTEL_R_SUPPRESS_SCOPES}
environment variable, or if the tracer was created before tracing
was turned on. The default \code{NULL} does not have this problem.}
}
\value{
The active span, an \link{otel_span} object, if any, or an invalid
span if there is no active span.
}
\description{
This is sometimes useful, to add additional attributes or links to the
currently active span.
}
\examples{
fun <- function() {
  otel::start_local_active_span("fun")
  spn <- otel::get_active_span()
  spn$set_attribute("key", "attribute-value")
}
fun()
}
