% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_active_span_context}
\alias{get_active_span_context}
\title{Returns the active span context}
\usage{
get_active_span_context(tracer = NULL)
}
\arguments{
\item{tracer}{Tracer object (\link{otel_tracer}) or tracer name to use.
If \code{NULL}, then otel uses an internal tracer. You usually do not
need to set this, the active span does not depend on the tracer.

Passing a tracer might give the wrong result. If it is a no-op
tracer, then the result is an invalid span (context), or no HTTP
headers, even if there is an active span. This happens if the
tracer's scope is turned off, e.g. via the \code{OTEL_R_SUPPRESS_SCOPES}
environment variable, or if the tracer was created before tracing
was turned on. The default \code{NULL} does not have this problem.}
}
\value{
The active span context, an \link{otel_span_context} object.
If there is no active span context, then an invalid span context is
returned, i.e. \code{spc$is_valid()} will be \code{FALSE} for the returned \code{spc}.
}
\description{
This is sometimes useful for logs or metrics, to associate
logging and metrics reporting with traces.
}
\details{
Note that logs and metrics instruments automatically use the current
span context, so often you don't need to call this function explicitly.
}
\examples{
fun <- function() {
  otel::start_local_active_span("fun")
  fun2()
}
fun2 <- function() {
  otel::log("Log message", span_context = otel::get_active_span_context())
}
fun()
}
