% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_current_span_context}
\alias{get_current_span_context}
\title{Return the current span context}
\usage{
get_current_span_context()
}
\value{
The current span context. If these is no current span context,
then an invalid span context is returned, i.e. \code{spc$is_valid()} will be
\code{FALSE} for the returned \code{spc}.
}
\description{
This is sometimes useful when writing loggers or meters, to associate
logging and metrics reporting with traces.
}
