% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_default_tracer}
\alias{get_default_tracer}
\title{Get a tracer from the default tracer provider}
\usage{
get_default_tracer(name = NULL)
}
\arguments{
\item{name}{Name of the new tracer. This is typically the name of the
package or project. Defaults to the name of the calling package,
or the name of the current working directory if not called from a
package.}
}
\value{
An OpenTelemetry tracer, an \code{otel_tracer} object.
}
\description{
Calls \code{\link[=get_default_tracer_provider]{get_default_tracer_provider()}} to get the default tracer
provider. Then calls its \verb{$get_tracer()} method to create a new tracer.
}
