% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_logger}
\alias{get_logger}
\title{Get a logger from the default logger provider}
\usage{
get_logger(name = NULL)
}
\arguments{
\item{name}{Name of the new tracer. If missing, then deduced automatically.}
}
\description{
Get a logger from the default logger provider
}
\seealso{
Other OpenTelemetry logging: 
\code{\link{is_logging}()},
\code{\link{log}()},
\code{\link{log_severity_levels}}
}
\concept{OpenTelemetry logging}
