% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_meter}
\alias{get_meter}
\title{Get a meter from the default meter provider}
\usage{
get_meter(name = NULL)
}
\arguments{
\item{name}{Name of the new meter. This is typically the name of the
package or project. Defaults to the name of the calling package,
or the name of the current working directory if not called from a
package.}
}
\description{
Get a meter from the default meter provider
}
