% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_tracer}
\alias{get_tracer}
\title{Get a tracer from the default tracer provider}
\usage{
get_tracer(name = NULL)
}
\arguments{
\item{name}{Name of the new tracer. If missing, then deduced automatically.}
}
\value{
An OpenTelemetry tracer, an \code{otel_tracer} object.
}
\description{
Calls \code{\link[=get_default_tracer_provider]{get_default_tracer_provider()}} to get the default tracer
provider. Then calls its \verb{$get_tracer()} method to create a new tracer.
}
\seealso{
Other OpenTelemetry tracing: 
\code{\link{extract_http_context}()},
\code{\link{get_active_span_context}()},
\code{\link{is_tracing}()},
\code{\link{start_session}()},
\code{\link{start_span}()}
}
\concept{OpenTelemetry tracing}
