% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{is_logging}
\alias{is_logging}
\title{Check whether OpenTelemetry logging is active}
\usage{
is_logging(logger = NULL)
}
\arguments{
\item{logger}{Logger object (\code{otel_logger}), or a logger name, the
instrumentation scope, to pass to \code{\link[=get_logger]{get_logger()}}.}
}
\value{
\code{TRUE} is OpenTelemetry logging is active, \code{FALSE} otherwise.
}
\description{
This is useful for avoiding computation when logging is inactive.
}
\details{
It calls \code{\link[=get_logger]{get_logger()}} with \code{name} and then it calls the logger's
\verb{$is_enabled()} method.
}
\seealso{
Other OpenTelemetry logging: 
\code{\link{get_logger}()},
\code{\link{log}()},
\code{\link{log_severity_levels}}
}
\concept{OpenTelemetry logging}
