% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{is_tracing}
\alias{is_tracing}
\title{Check whether OpenTelemetry tracing is active}
\usage{
is_tracing(tracer = NULL)
}
\arguments{
\item{tracer}{Tracer object (\code{otel_tracer}), or a tracer name, the
instrumentation scope, to pass to \code{\link[=get_tracer]{get_tracer()}}.}
}
\value{
\code{TRUE} is OpenTelemetry tracing is active, \code{FALSE} otherwise.
}
\description{
This is useful for avoiding computation when tracing is inactive.
}
\details{
It calls \code{\link[=get_tracer]{get_tracer()}} with \code{name} and then it calls the tracer's
\verb{$is_enabled()} method.
}
\concept{OpenTelemetry tracing
TODO}
