% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{local_active_span}
\alias{local_active_span}
\title{Activate an OpenTelemetry span for an R scope}
\usage{
local_active_span(span, end_on_exit = FALSE, activation_scope = parent.frame())
}
\arguments{
\item{span}{The OpenTelemetry span to activate.}

\item{end_on_exit}{Whether to end the span when exiting the activation
scope.}

\item{activation_scope}{The scope to activate the span for, defaults to
the caller frame.}
}
\description{
Activate an OpenTelemetry span for an R scope
}
