% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{local_session}
\alias{local_session}
\title{Activate an OpenTelemetry session span for a frame}
\usage{
local_session(session, session_scope = parent.frame())
}
\arguments{
\item{session}{The OpenTelemetry session span to activate.}

\item{session_scope}{The frame to activate it for, defaults to the
caller frame.}
}
\description{
Activate an OpenTelemetry session span for a frame
}
