% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger-provider-noop.R
\docType{data}
\name{logger_provider_noop}
\alias{logger_provider_noop}
\title{No-op logger provider}
\format{
An object of class \code{list} of length 1.
}
\usage{
logger_provider_noop
}
\description{
No-op logger provider
}
\keyword{internal}
