% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otel-package.R
\docType{package}
\name{otel-package}
\alias{otel}
\alias{otel-package}
\title{otel: R Implementation of 'OpenTelemetry', exporters}
\description{
High-quality, ubiquitous, and portable telemetry to enable effective observability. 'OpenTelemetry' is a collection of tools, 'APIs', and 'SDKs' used to instrument, generate, collect, and export telemetry data (metrics, logs, and traces) for analysis in order to understand your software's performance and behavior. Use this package as a dependency if you want to instrument your R package for 'OpenTelemetry'.
}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

}
\keyword{internal}
