% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{otel_meter}
\alias{otel_meter}
\title{TODO}
\description{
TODO
}
\seealso{
Other OpenTelemetry metrics: 
\code{\link{counter_add}()},
\code{\link{gauge_record}()},
\code{\link{get_meter}()},
\code{\link{histogram_record}()},
\code{\link{is_measuring_enabled}()},
\code{\link{up_down_counter_add}()}
}
\concept{OpenTelemetry metrics}
