% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otel-package.R
\name{otel_tracer_name}
\alias{otel_tracer_name}
\title{Default tracer, meter, logger name for an R package}
\description{
The otel R package uses the \code{otel_tracer_name} variable from your
package when it deducts the tracer (and meter and logger) name to use
if \code{\link[=start_span]{start_span()}} (etc.) is called from your package.
}
\details{
We suggest that you set this internal variable in your package to avoid
having to specify the tracer name for every \code{\link[=start_span]{start_span()}} call.

The \href{https://opentelemetry.io/docs/specs/otel/trace/api/#get-a-tracer}{OpenTelemetry specification} recommends using a tracer name that identifies the instrumentation
scope, i.e. your package.

Some tips on choosing the tracer name:
\itemize{
\item If your R package can be associated with a URL, you can use the
"reverse" of that URL. E.g. since the callr package's online manual
is at https://callr.r-lib.org, it can use \code{org.r-lib.callr}.
\item If your R package belongs to your company, you can use the "reverse"
of the company URL, possibly with an additional prefix. E.g. for the
shiny R package by Posit, \code{co.posit.r-package.shiny} seems like a
good name.
\item If you don't set \code{otel_tracer_name}, then otel will use
\verb{r.package.<package-name>} as the tracer name.
}
}
