% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{start_session}
\alias{start_session}
\title{Start a new OpenTelemetry session span, using the default tracer}
\usage{
start_session(name = NULL, ...)
}
\arguments{
\item{name}{Name of the span.}

\item{...}{Additional arguments are passed to the default tracer's
\code{start_session()} method.}
}
\description{
A session span is special because it has its own context stack, so it
behaves like a separate process. You can switch to this "process" by
calling the span's \verb{$activate_session()} method, and switch back to
context stack of the main process by calling the \verb{$deactivate_session}
method.
}
\seealso{
Other OpenTelemetry tracing: 
\code{\link{extract_http_context}()},
\code{\link{get_active_span_context}()},
\code{\link{get_tracer}()},
\code{\link{is_tracing}()},
\code{\link{start_span}()}
}
\concept{OpenTelemetry tracing}
