% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{start_span}
\alias{start_span}
\title{Start a new OpenTelemetry span, using the default tracer}
\usage{
start_span(
  name = NULL,
  tracer_name = NULL,
  attributes = NULL,
  links = NULL,
  options = NULL,
  scope = parent.frame(),
  activation_scope = parent.frame()
)
}
\arguments{
\item{name}{Name of the span. If not specified it will be \code{"<NA>"}.}

\item{tracer_name}{The name of the tracer to use, see \code{\link[=get_tracer]{get_tracer()}}.}

\item{attributes}{Span attributes. You may use \code{\link[=as_attributes]{as_attributes()}} to
convert R objects to OpenTelemetry attributes. OpenTelemetry supports
the following R types as attributes:
character, logical, double, integer.}

\item{links}{TODO}

\item{options}{TODO}

\item{scope}{TODO}

\item{activation_scope}{TODO}
}
\value{
The new OpenTelemetry span object (of class \code{otel_span}),
invisibly. See \link{otel_span} for information about the returned object.
}
\description{
A span represents a unit of work or operation. Spans are the building
blocks of traces. \code{start_span()} is typically called at the beginning
of a function or the beginning of some larger operation, e.g. an HTTP
request.
}
\details{
Calls \code{\link[=get_tracer]{get_tracer()}} with \code{tracer_name} as the tracer name, and then
calls the \verb{$start_span()} method of the tracer with the other arguments.
\subsection{Lifetime of a span}{

TODO
}

\subsection{The active span}{

TODO
}

\subsection{Span attributes}{

TODO
}

\subsection{Span links}{

TODO
}

\subsection{Span options}{

TODO
}
}
\seealso{
Other OpenTelemetry tracing: 
\code{\link{extract_http_context}()},
\code{\link{get_active_span_context}()},
\code{\link{get_tracer}()},
\code{\link{is_tracing}()}
}
\concept{OpenTelemetry tracing}
