% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-provider-noop.R
\docType{data}
\name{tracer_provider_noop}
\alias{tracer_provider_noop}
\title{No-op tracer provider}
\description{
This is the tracer provider otel uses when tracing is disabled.
}
\details{
All methods are no-ops or return objects that are also no-ops.
}
\seealso{
Other low level trace API: 
\code{\link{get_active_span_context}()},
\code{\link{get_default_tracer_provider}()},
\code{\link{otel_span}},
\code{\link{otel_span_context}},
\code{\link{otel_tracer}},
\code{\link{otel_tracer_provider}}
}
\concept{low level trace API}
\keyword{datasets}
