% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{tracing}
\alias{tracing}
\alias{invalid_trace_id}
\alias{invalid_span_id}
\alias{span_kinds}
\alias{span_status_codes}
\title{OpenTelemetry tracing constants}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 5.

An object of class \code{character} of length 3.
}
\usage{
invalid_trace_id

invalid_span_id

span_kinds

span_status_codes
}
\description{
Various constants related OpenTelemetry tracing.
}
\details{
\code{invalid_trace_id} is a string scalar, an invalid trace id.

\code{invalid_span_id} is a string scalar, an invalid span id.

\code{span_kinds} is a character vector listing all possible span kinds.

\code{span_status_codes} is a character vector listing all possible span
status codes.
}
\keyword{datasets}
