% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{with_session}
\alias{with_session}
\title{Evaluate R code with an active OpenTelemetry session span}
\usage{
with_session(session, expr)
}
\arguments{
\item{session}{The OpenTelemetry session span to activate.}

\item{expr}{R expression to evaluate}
}
\value{
The return value of \code{expr}.
}
\description{
Evaluate R code with an active OpenTelemetry session span
}
