% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{counter_add}
\alias{counter_add}
\title{Increase an OpenTelemetry counter using the default meter}
\usage{
counter_add(name, value = 1L, attributes = NULL, context = NULL)
}
\arguments{
\item{name}{Name of the counter.}

\item{value}{Value to add to the counter, defaults to 1.}

\item{attributes}{Additional attributes to add.}

\item{context}{Span context. If missing the active context is used,
if any.}
}
\value{
The counter object, invisibly.
}
\description{
Increase an OpenTelemetry counter using the default meter
}
\seealso{
Other OpenTelemetry metrics instruments: 
\code{\link{gauge_record}()},
\code{\link{histogram_record}()},
\code{\link{up_down_counter_add}()}

Other OpenTelemetry metrics: 
\code{\link{gauge_record}()},
\code{\link{get_meter}()},
\code{\link{histogram_record}()},
\code{\link{is_measuring}()},
\code{\link{up_down_counter_add}()}
}
\concept{OpenTelemetry metrics}
\concept{OpenTelemetry metrics instruments}
