% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_tracer_name}
\alias{default_tracer_name}
\title{Default tracer (and meter and logger) name for an R package}
\usage{
default_tracer_name()
}
\description{
Default tracer (and meter and logger) name for an R package
}
