% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{end_span}
\alias{end_span}
\title{End an OpenTelemetry span}
\usage{
end_span(span)
}
\arguments{
\item{span}{The span to end.}
}
\value{
\code{NULL}.
}
\description{
You must end every span, by calling \code{end_span}, or using the \code{end_on_exit}
argument of \code{\link[=local_active_span]{local_active_span()}} or \code{\link[=with_active_span]{with_active_span()}}.
}
