% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{extract_http_context}
\alias{extract_http_context}
\title{Extract a span context from HTTP headers received from a client}
\usage{
extract_http_context(headers)
}
\arguments{
\item{headers}{A named list with one or two strings: \code{traceparent} is
mandatory, and \code{tracestate} is optional.}
}
\description{
The return value can be used as the \code{parent} option when starting
a span.
}
\seealso{
Other OpenTelemetry tracing: 
\code{\link{get_active_span_context}()},
\code{\link{get_tracer}()},
\code{\link{start_local_active_span}()},
\code{\link{start_span}()}
}
\concept{OpenTelemetry tracing}
