% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gauge_record}
\alias{gauge_record}
\title{Record a value of an OpenTelemetry gauge using the default
meter}
\usage{
gauge_record(
  name,
  value,
  attributes = NULL,
  context = NULL,
  session = NULL,
  session_scope = parent.frame()
)
}
\arguments{
\item{name}{Name of the gauge}

\item{value}{Value to record.}

\item{attributes}{Additional attributes to add.}

\item{context}{Span context. If missing the active context is used,
if any.}

\item{session}{Optionally a session span to activate before updating
the counter.}

\item{session_scope}{The session span will be automatically deactivated
after this scope.}
}
\value{
The gauge object, invisibly.
}
\description{
Record a value of an OpenTelemetry gauge using the default
meter
}
\seealso{
Other OpenTelemetry metrics instruments: 
\code{\link{counter_add}()},
\code{\link{histogram_record}()},
\code{\link{up_down_counter_add}()}

Other OpenTelemetry metrics: 
\code{\link{counter_add}()},
\code{\link{get_meter}()},
\code{\link{histogram_record}()},
\code{\link{is_measuring}()},
\code{\link{up_down_counter_add}()}
}
\concept{OpenTelemetry metrics}
\concept{OpenTelemetry metrics instruments}
