% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_active_span_context}
\alias{get_active_span_context}
\title{Returns the active span context}
\usage{
get_active_span_context()
}
\value{
The active span context. If these is no active span context,
then an invalid span context is returned, i.e. \code{spc$is_valid()} will be
\code{FALSE} for the returned \code{spc}.
}
\description{
This is sometimes useful when writing loggers or meters, to associate
logging and metrics reporting with traces.
}
\seealso{
Other OpenTelemetry tracing: 
\code{\link{extract_http_context}()},
\code{\link{get_tracer}()},
\code{\link{is_tracing}()},
\code{\link{start_session}()},
\code{\link{start_span}()}
}
\concept{OpenTelemetry tracing}
