% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_logger}
\alias{get_logger}
\title{Get a logger from the default logger provider}
\usage{
get_logger(name = NULL)
}
\arguments{
\item{name}{Name of the new logger. This is typically the name of the
package or project. Defaults to the name of the calling package,
or the name of the current working directory if not called from a
package.}
}
\description{
Get a logger from the default logger provider
}
