% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_meter}
\alias{get_meter}
\title{Get a meter from the default meter provider}
\usage{
get_meter(name = NULL)
}
\arguments{
\item{name}{Name of the new tracer. If missing, then deduced automatically.}
}
\description{
Get a meter from the default meter provider
}
\seealso{
Other OpenTelemetry metrics: 
\code{\link{counter_add}()},
\code{\link{gauge_record}()},
\code{\link{histogram_record}()},
\code{\link{is_measuring}()},
\code{\link{up_down_counter_add}()}
}
\concept{OpenTelemetry metrics}
