% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{histogram_record}
\alias{histogram_record}
\title{Record a value of an OpenTelemetry histogram using the default
meter}
\usage{
histogram_record(name, value, attributes = NULL, context = NULL)
}
\arguments{
\item{name}{Name of the histogram.}

\item{value}{Value to record.}

\item{attributes}{Additional attributes to add.}

\item{context}{Span context. If missing the active context is used,
if any.}
}
\value{
The histogram object, invisibly.
}
\description{
Record a value of an OpenTelemetry histogram using the default
meter
}
\seealso{
Other OpenTelemetry metrics instruments: 
\code{\link{counter_add}()},
\code{\link{gauge_record}()},
\code{\link{up_down_counter_add}()}
}
\concept{OpenTelemetry metrics instruments}
