% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{is_logging}
\alias{is_logging}
\title{Check whether OpenTelemetry logging is active}
\usage{
is_logging()
}
\value{
\code{TRUE} is OpenTelemetry logging is active, \code{FALSE} otherwise.
}
\description{
Check whether OpenTelemetry logging is active
}
\seealso{
Other OpenTelemetry logging: 
\code{\link{get_logger}()},
\code{\link{log}()},
\code{\link{log_severity_levels}}
}
\concept{OpenTelemetry logging}
