% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{is_tracing}
\alias{is_tracing}
\title{Check whether OpenTelemetry tracing is active}
\usage{
is_tracing()
}
\value{
\code{TRUE} is OpenTelemetry tracing is active, \code{FALSE} otherwise.
}
\description{
Check whether OpenTelemetry tracing is active
}
\seealso{
Other OpenTelemetry tracing: 
\code{\link{extract_http_context}()},
\code{\link{get_active_span_context}()},
\code{\link{get_tracer}()},
\code{\link{start_session}()},
\code{\link{start_span}()}
}
\concept{OpenTelemetry tracing}
