% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{local_active_span}
\alias{local_active_span}
\title{Activate an OpenTelemetry span for an R scope}
\usage{
local_active_span(span, activation_scope = parent.frame())
}
\arguments{
\item{span}{The OpenTelemetry span to activate.}

\item{activation_scope}{The scope to activate it for, defaults to the
caller frame.}
}
\description{
Activate an OpenTelemetry span for an R scope
}
