% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{log}
\alias{log}
\title{Log an OpenTelemetry log message, using the default logger}
\usage{
log(msg, ..., severity = "info", .envir = parent.frame())
}
\arguments{
\item{msg}{Log message, may contain R expressions to evaluate within
braces.}

\item{...}{Additional arguments are passed to the \verb{$log()} method of
the default logger.}

\item{severity}{Log severity, a string, one of
"1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24".}

\item{.envir}{Environment to evaluate the interpolated  expressions of
the log message in.}
}
\value{
The logger, invisibly.
}
\description{
Log an OpenTelemetry log message, using the default logger
}
\seealso{
Other OpenTelemetry logging: 
\code{\link{get_logger}()},
\code{\link{is_logging}()},
\code{\link{log_severity_levels}}
}
\concept{OpenTelemetry logging}
