% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{log}
\alias{log}
\alias{log_trace}
\alias{log_debug}
\alias{log_info}
\alias{log_warn}
\alias{log_error}
\alias{log_fatal}
\title{Log an OpenTelemetry log message, using the default logger}
\usage{
log(msg, ..., severity = "info", .envir = parent.frame())

log_trace(msg, ..., .envir = parent.frame())

log_debug(msg, ..., .envir = parent.frame())

log_info(msg, ..., .envir = parent.frame())

log_warn(msg, ..., .envir = parent.frame())

log_error(msg, ..., .envir = parent.frame())

log_fatal(msg, ..., .envir = parent.frame())
}
\arguments{
\item{msg}{Log message, may contain R expressions to evaluate within
braces.}

\item{...}{Additional arguments are passed to the \verb{$log()} method of
the default logger.}

\item{severity}{Log severity, a string, one of
"1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24".}

\item{.envir}{Environment to evaluate the interpolated  expressions of
the log message in.}
}
\value{
The logger, invisibly.
}
\description{
Log an OpenTelemetry log message, using the default logger
}
\details{
\code{log_trace()} is the same as \code{log()} with \code{severity_level}
"trace".

\code{log_debug()} is the same as \code{log()} with \code{severity_level}
"debug".

\code{log_info()} is the same as \code{log()} with \code{severity_level}
"info".

\code{log_warn()} is the same as \code{log()} with \code{severity_level}
"warn".

\code{log_error()} is the same as \code{log()} with \code{severity_level}
"error".

\code{log_fatal()} is the same as \code{log()} with \code{severity_level}
"fatal".
}
\seealso{
Other OpenTelemetry logging: 
\code{\link{get_logger}()},
\code{\link{is_logging}()},
\code{\link{log_severity_levels}}
}
\concept{OpenTelemetry logging}
