% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{log_severity_levels}
\alias{log_severity_levels}
\title{OpenTelemetry log severity levels}
\value{
Not applicable.
}
\description{
A named integer vector, the severity levels in numeric form.
The names are the severity levels in text form. otel functions accept
both forms as severity levels, but the text form is more readable.
}
\examples{
log_severity_levels
}
\seealso{
Other OpenTelemetry logs API: 
\code{\link{is_logging_enabled}()},
\code{\link{log}()}
}
\concept{OpenTelemetry logs API}
\keyword{datasets}
