% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger-provider-noop.R
\docType{data}
\name{logger_provider_noop}
\alias{logger_provider_noop}
\title{No-op logger provider}
\description{
This is the logger provider (\link{otel_logger_provider}) otel uses when
logging is disabled.
}
\details{
All methods are no-ops or return objects that are also no-ops.
}
\seealso{
Other low level logs API: 
\code{\link{otel_logger}},
\code{\link{otel_logger_provider}}
}
\concept{low level logs API}
\keyword{internal}
