% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger-provider-noop.R
\docType{data}
\name{logger_provider_noop}
\alias{logger_provider_noop}
\title{No-op logger provider}
\value{
Not applicable.
}
\description{
This is the logger provider (\link{otel_logger_provider}) otel uses when
logging is disabled.
}
\details{
All methods are no-ops or return objects that are also no-ops.
}
\examples{
logger_provider_noop$new()
}
\seealso{
Other low level logs API: 
\code{\link{get_default_logger_provider}()},
\code{\link{get_logger}()},
\code{\link{otel_logger}},
\code{\link{otel_logger_provider}}
}
\concept{low level logs API}
\keyword{internal}
