% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter-provider-noop.R
\docType{data}
\name{meter_provider_noop}
\alias{meter_provider_noop}
\title{No-op Meter Provider}
\value{
Not applicable.
}
\description{
This is the meter provider (\link{otel_meter_provider}) otel uses when
metrics collection is disabled.
}
\details{
All methods are no-ops or return objects that are also no-ops.
}
\examples{
meter_provider_noop$new()
}
\seealso{
Other low level metrics API: 
\code{\link{get_default_meter_provider}()},
\code{\link{get_meter}()},
\code{\link{otel_counter}},
\code{\link{otel_gauge}},
\code{\link{otel_histogram}},
\code{\link{otel_meter}},
\code{\link{otel_meter_provider}},
\code{\link{otel_up_down_counter}}
}
\concept{low level metrics API}
\keyword{datasets}
