% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-provider-noop.R
\name{otel_tracer_provider}
\alias{otel_tracer_provider}
\title{OpenTelemetry tracer provider objects}
\description{
The tracer provider defines how traces are exported when collecting
telemetry data. It is unlikely that you'd need to use tracer provider
objects directly.
}
\details{
\subsection{Implementations}{

Note that this list is updated manually and may be incomplete.
\itemize{
\item \link{tracer_provider_noop}: No-op tracer provider, used when no traces are
emitted.
\item \link[otelsdk:tracer_provider_file]{otelsdk::tracer_provider_file}: Save traces to a JSONL file.
\item \link[otelsdk:tracer_provider_http]{otelsdk::tracer_provider_http}: Send traces to a collector over
HTTP/OTLP.
\item \link[otelsdk:tracer_provider_memory]{otelsdk::tracer_provider_memory}: Collect emitted traces in memory.
For testing.
\item \link[otelsdk:tracer_provider_stdstream]{otelsdk::tracer_provider_stdstream}: Write traces to standard output
or error or to a file. Primarily for debugging.
}
}

\subsection{Methods}{
\subsection{Method: \code{get_tracer()}}{
\subsection{Description}{

Get or create a new tracer object.
}

\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{$get_tracer(
  name = NULL,
  version = NULL,
  schema_url = NULL,
  attributes = NULL
)
}\if{html}{\out{</div>}}
}

\subsection{Arguments:}{
\itemize{
\item \code{name}: Tracer name, see \code{\link[=get_tracer]{get_tracer()}}.
\item \code{version}: Optional. Specifies the version of the instrumentation
scope if the scope has a version (e.g. R package version).
Example value: \code{"1.0.0"}.
\item \code{schema_url}: Optional. Specifies the Schema URL that should be
recorded in the emitted telemetry.
\item \code{attributes}: Optional. Specifies the instrumentation scope
attributes to associate with emitted telemetry.
}
}

\subsection{See also}{

\code{\link[=get_tracer]{get_tracer()}}.
}

}

\subsection{Method: \code{flush()}}{
\subsection{Description}{

Force any buffered spans to flush. Tracer providers might not implement
this method.
}

\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{$flush()
}\if{html}{\out{</div>}}
}

}

}
}
