% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pack_http_context}
\alias{pack_http_context}
\title{Pack the currently active span context into standard HTTP OpenTelemetry
headers}
\usage{
pack_http_context()
}
\value{
A named character vector, with lovercase names.
}
\description{
The returned headers can be sent over HTTP, or set as environment
variables for subprocesses.
}
