% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{start_session}
\alias{start_session}
\title{Start a new OpenTelemetry session span, using the default tracer}
\usage{
start_session(name = NULL, ..., session_scope = parent.frame())
}
\arguments{
\item{name}{Name of the span.}

\item{..., session_scope}{Additional arguments are passed to the default
tracer's \code{start_session()} method.}
}
\description{
Session spans are for instrumenting concurrent code in R.
}
\details{
A session span does not end at the end of an R function, but has to
be ended explicitly, using its \verb{$end()} method.

A session span can be activated using \code{\link[=with_session]{with_session()}} or
\code{\link[=local_session]{local_session()}}.
}
\seealso{
Other OpenTelemetry tracing: 
\code{\link{extract_http_context}()},
\code{\link{get_active_span_context}()},
\code{\link{get_tracer}()},
\code{\link{is_tracing}()},
\code{\link{start_span}()}
}
\concept{OpenTelemetry tracing}
