% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{up_down_counter_add}
\alias{up_down_counter_add}
\title{Increase or decrease an OpenTelemetry up-down counter using the default
meter}
\usage{
up_down_counter_add(name, value = 1L, attributes = NULL, context = NULL)
}
\arguments{
\item{name}{Name of the up-down counter.}

\item{value}{Value to add to or subtract from the counter, defaults
to 1.}

\item{attributes}{Additional attributes to add.}

\item{context}{Span context. If missing the active context is used,
if any.}
}
\value{
The up-down counter object, invisibly.
}
\description{
Increase or decrease an OpenTelemetry up-down counter using the default
meter
}
\seealso{
Other OpenTelemetry metrics instruments: 
\code{\link{counter_add}()},
\code{\link{gauge_record}()},
\code{\link{histogram_record}()}
}
\concept{OpenTelemetry metrics instruments}
